+++ src/WhereConstraints/WhereConstraintsServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Nuwave\Lighthouse\WhereConstraints;
 
+use GraphQL\Language\AST\InputObjectTypeDefinitionNode;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Contracts\Events\Dispatcher;
 use Nuwave\Lighthouse\Events\ManipulateAST;
@@ -29,16 +30,11 @@ class WhereConstraintsServiceProvider extends ServiceProvider
             function (ManipulateAST $manipulateAST): void {
                 $manipulateAST->documentAST
                     ->setTypeDefinition(
-                        PartialParser::inputObjectTypeDefinition('
-                            input WhereConstraints {
-                                column: String
-                                operator: Operator = EQ
-                                value: Mixed
-                                AND: [WhereConstraints!]
-                                OR: [WhereConstraints!]
-                                NOT: [WhereConstraints!]
-                            }
-                        ')
+                        static::createWhereConstraintsInputType(
+                            'WhereConstraints',
+                            'Dynamic WHERE constraints for queries.',
+                            'String'
+                        )
                     )
                     ->setTypeDefinition(
                         PartialParser::scalarTypeDefinition('
@@ -47,5 +43,19 @@ class WhereConstraintsServiceProvider extends ServiceProvider
                     );
             }
         );
+    }
+
+    public static function createWhereConstraintsInputType(string $name, string $description, string $columnType): InputObjectTypeDefinitionNode
+    {
+        return PartialParser::inputObjectTypeDefinition("
+            input $name {
+                column: $columnType
+                operator: Operator = EQ
+                value: Mixed
+                AND: [$name!]
+                OR: [$name!]
+                NOT: [$name!]
+            }
+        ");
     }
 }