+++ tests/Integration/IntrospectionTest.php
@@ -10,87 +10,6 @@ use Illuminate\Foundation\Testing\TestResponse;
 
 class IntrospectionTest extends TestCase
 {
-    /**
-     * @see https://gist.github.com/craigbeck/b90915d49fda19d5b2b17ead14dcd6da
-     */
-    const INTROSPECTION_QUERY = /* @lang GraphQL */
-        <<<'GRAPHQL'
-  query IntrospectionQuery {
-    __schema {
-      queryType { name }
-      mutationType { name }
-      subscriptionType { name }
-      types {
-        ...FullType
-      }
-      directives {
-        name
-        description
-        args {
-          ...InputValue
-        }
-        locations
-      }
-    }
-  }
-
-  fragment FullType on __Type {
-    kind
-    name
-    description
-    fields(includeDeprecated: true) {
-      name
-      description
-      args {
-        ...InputValue
-      }
-      type {
-        ...TypeRef
-      }
-      isDeprecated
-      deprecationReason
-    }
-    inputFields {
-      ...InputValue
-    }
-    interfaces {
-      ...TypeRef
-    }
-    enumValues(includeDeprecated: true) {
-      name
-      description
-      isDeprecated
-      deprecationReason
-    }
-    possibleTypes {
-      ...TypeRef
-    }
-  }
-
-  fragment InputValue on __InputValue {
-    name
-    description
-    type { ...TypeRef }
-    defaultValue
-  }
-
-  fragment TypeRef on __Type {
-    kind
-    name
-    ofType {
-      kind
-      name
-      ofType {
-        kind
-        name
-        ofType {
-          kind
-          name
-        }
-      }
-    }
-  }
-GRAPHQL;
 
     /**
      * @var \Nuwave\Lighthouse\Schema\TypeRegistry
@@ -114,21 +33,22 @@ GRAPHQL;
      */
     public function itFindsTypesFromSchema(): void
     {
-        $this->introspect('
+        $this->schema = '
         type Foo {
             bar: Int
         }        
-        '.$this->placeholderQuery()
-        );
+        '.$this->placeholderQuery();
+
+        $introspection = $this->introspect();
 
         $this->assertTrue(
-            $this->isTypeNamePresent('Foo')
+            $this->isTypeNamePresent($introspection, 'Foo')
         );
         $this->assertTrue(
-            $this->isTypeNamePresent('Query')
+            $this->isTypeNamePresent($introspection, 'Query')
         );
         $this->assertFalse(
-            $this->isTypeNamePresent('Bar')
+            $this->isTypeNamePresent($introspection, 'Bar')
         );
     }
 
@@ -137,29 +57,19 @@ GRAPHQL;
      */
     public function itFindsManuallyRegisteredTypes(): void
     {
+        $this->schema = $this->placeholderQuery();
         $this->typeRegistry->register(
             new Email()
         );
 
-        $this->introspect(
-            $this->placeholderQuery()
-        );
-
         $this->assertTrue(
-            $this->isTypeNamePresent('Email')
+            $this->isTypeNamePresent($this->introspect(), 'Email')
         );
     }
 
-    protected function introspect(string $schema): void
+    protected function isTypeNamePresent(TestResponse $introspection, string $typeName): bool
     {
-        $this->schema = $schema;
-
-        $this->introspectionResult = $this->graphQL(self::INTROSPECTION_QUERY);
-    }
-
-    protected function isTypeNamePresent(string $typeName): bool
-    {
-        return (new Collection($this->introspectionResult->jsonGet('data.__schema.types')))
+        return (new Collection($introspection->jsonGet('data.__schema.types')))
             ->contains(function (array $type) use ($typeName): bool {
                 return $type['name'] === $typeName;
             });