+++ tests/Integration/WhereConstraints/WhereConstraintsDirectiveTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Integration\WhereConstraints;
 
+use Illuminate\Support\Arr;
 use Tests\DBTestCase;
 use Tests\Utils\Models\User;
 use Nuwave\Lighthouse\WhereConstraints\WhereConstraintsDirective;
@@ -18,6 +19,9 @@ class WhereConstraintsDirectiveTest extends DBTestCase
     
     type Query {
         users(where: WhereConstraints @whereConstraints): [User!]! @all
+        whitelistedColumns(
+            where: WhereConstraints @whereConstraints(columns: ["id", "camelCase"])
+        ): [User!]! @all
     }
     
     enum Operator {
@@ -281,5 +285,59 @@ class WhereConstraintsDirectiveTest extends DBTestCase
         ')->assertJsonFragment([
             'message' => WhereConstraintsDirective::missingValueForColumn('no_value'),
         ]);
+    }
+
+    /**
+     * @test
+     */
+    public function itOnlyAllowsWhitelistedColumns(): void
+    {
+        factory(User::class)->create();
+
+        $this->graphQL('
+        {
+            whitelistedColumns(
+                where: {
+                    column: ID
+                    value: 1
+                }
+            ) {
+                id
+            }
+        }
+        ')->assertJson([
+            'data' => [
+                'whitelistedColumns' => [
+                    [
+
+                    'id' => 1
+                    ]
+                ]
+            ]
+        ]);
+
+        $types = $this->introspect()->jsonGet('data.__schema.types');
+
+        $expectedEnumName = 'WhitelistedColumnsWhereColumn';
+        $enum = Arr::first($types, function(array $type) use ($expectedEnumName): bool {
+            return $type['name'] === $expectedEnumName;
+        });
+
+        $this->assertArraySubset(
+            [
+                'kind' => 'ENUM',
+                'name' => $expectedEnumName,
+                'description' => 'Allowed column names for the `where` argument on the query `whitelistedColumns`.',
+                'enumValues' => [
+                    [
+                        'name' => 'ID',
+                    ],
+                    [
+                        'name' => 'CAMEL_CASE',
+                    ],
+                ]
+            ],
+            $enum
+        );
     }
 }